feat: basic HBase client for Scala
This commit implements a basic HBase client for Scala. It is built as a
wrapper around the HBase Java client library. HBase operations are
modelled as functions instead of classes. Salsa Beach allows the
developer to represent HBase row keys as any type and translates to
byte arrays and back using a Type Class approach.